Implement utility function for terminal invokation, improvements
- Now the installer uses annutility function to invoke the terminal.
This way the terminal is not executed in a detached process.
- Echo out the script name before executing it
- Echo message out if the script finished. In the case
of konsole we use the --noclose option, this means the window
stays open after the script finished. By echoing out
a message we tell the user what to do now.
- If there are multiple scripts that start with the prefix
we prefer the one which has an exact match with the basename.